+++ test/test_fiber.rb
@@ -142,7 +142,7 @@ class FiberJoinTest < UMBaseTest
 end
 
 class WaitFibersTest < UMBaseTest
-  def test_await_fibers
+  def test_await
     q = UM::Queue.new
     x = nil
 
@@ -163,19 +163,19 @@ class WaitFibersTest < UMBaseTest
       machine.push q, 2
     end
 
-    res = machine.await_fibers([f])
+    res = machine.await([f])
     assert_equal 0, x
     assert_equal 3, machine.shift(q)
     assert_equal 1, res
 
     done = nil
     f = machine.spin { machine.snooze; done = true }
-    res = machine.await_fibers(f)
+    res = machine.await(f)
     assert done
     assert_equal 1, res
   end
 
-  def test_await_fibers_multiple
+  def test_await_multiple
     f1 = machine.spin do
       :foo
     end
@@ -189,14 +189,14 @@ class WaitFibersTest < UMBaseTest
       :baz
     end
 
-    res = machine.await_fibers([f1, f2, f3])
+    res = machine.await([f1, f2, f3])
     assert_equal 3, res
     assert_equal true, f1.done?
     assert_equal true, f2.done?
     assert_equal true, f3.done?
   end
 
-  def test_await_fibers_multiple_splat
+  def test_await_multiple_splat
     f1 = machine.spin do
       :foo
     end
@@ -210,14 +210,14 @@ class WaitFibersTest < UMBaseTest
       :baz
     end
 
-    res = machine.await_fibers(f1, f2, f3)
+    res = machine.await(f1, f2, f3)
     assert_equal 3, res
     assert_equal true, f1.done?
     assert_equal true, f2.done?
     assert_equal true, f3.done?
   end
 
-  def test_await_fibers_cross_thread
+  def test_await_cross_thread
     q = UM::Queue.new
 
     t2 = Thread.new do
@@ -227,32 +227,32 @@ class WaitFibersTest < UMBaseTest
         m2.snooze
         :foo
       end
-      m2.await_fibers(f)
+      m2.await(f)
     end
 
     f = machine.shift(q)
     assert_kind_of Fiber, f
-    res = machine.await_fibers(f)
+    res = machine.await(f)
     assert_equal 1, res
   ensure
     t2.join
   end
 
-  def test_await_fibers_with_exception
+  def test_await_with_exception
     f = machine.spin do
       raise "Foobar"
     end
 
-    res = machine.await_fibers(f)
+    res = machine.await(f)
     assert_equal 1, res
   end
 
-  def test_await_fibers_terminate
+  def test_await_terminate
     f1 = machine.spin { machine.sleep(1) }
     f2 = machine.spin { machine.sleep(1) }
     done = false
     a = machine.spin do
-      machine.await_fibers([f1, f2])
+      machine.await([f1, f2])
     rescue UM::Terminate
       done = true
     end